+++ ModAudio/_Thunderstore/CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.0] - 2025-Mar-29
+
+### Added
+
+- Added an effect option "overlays_ignore_restarts" for `__routes.txt` routes (`originalClip @ overlayClip ~ overlays_ignore_restarts : true`)
+  - Settings this as true will make it so that an overlay sound is played only once if an audio source is restarted multiple times.
+
+### Changed
+
+- Slightly improved load times by skipping loading audio data for audio packs that are disabled
+- Mod assembly now uses an embedded PDB instead of a separate PDB file for debugging
+
+### Fixed
+
+- Fixed some playOnAwake sources being logged an extra time in the console
+- Fixed pitch and volume being tracked and applied incorrectly for cast effect sounds
+
 ## [2.0.1] - 2025-Mar-19
 
 ### Changed